Llamacloud_ocr_api/Principal/requirements.txt

python-dotenv>=0.9.9
fastapi[standard]>=0.117.1
llama-cloud-services>=0.6.68
numpy>=2.3.3
pandas>=2.3.2
pydantic>=2.11.9
uvicorn>=0.36.0
